Reorganize the CLI
Reorganizes the CLI around Typer sub-commands. The available
commands are compute, export, import, prepare, and run.

Fixes PeopleForBikes/brokenspoke-analyzer/issues/247
